requirements.txt
Flaskgunicornplaywrightpandasgspreadoauth2clientPillowopenpyxltwocaptcha-python